refactor display fucntion to be a module, import dislpay module
